requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
pydantic-settings
google-cloud-secret-manager
python-multipart
aiofiles